Refactor manager reducer
