+++ app/reducers/managerReducer/index.ts
@@ -1,0 +1,2 @@
+export { default, emptyState } from './managerReducer';
+export type { ManagerData } from './managerReducer';
