+++ app/reducers/managerReducer/tenantsReducer.ts
@@ -1,9 +1,16 @@
-// @ts-nocheck File not migrated fully to TS
-
+import type { Reducer } from 'redux';
 import _ from 'lodash';
-import * as types from '../actions/types';
+import * as types from '../../actions/types';
 
-const tenants = (state = {}, action) => {
+export interface TenantsObject {
+    isFetching?: boolean;
+    items?: string[];
+    selected?: string;
+    lastUpdated?: string;
+    error?: string;
+}
+
+const tenants: Reducer<TenantsObject> = (state = {}, action) => {
     let selectedTenant;
     switch (action.type) {
         case types.REQ_TENANTS:
