+++ test/jest/reducers/managerReducer.test.ts
@@ -8,7 +8,7 @@ import { applyMiddleware, createStore } from 'redux';
 import { getManagerData, login, logout } from 'actions/managers';
 import * as types from 'actions/types';
 import managerReducer, { emptyState } from 'reducers/managerReducer';
-import licenseReducer from 'reducers/licenseReducer';
+import licenseReducer from 'reducers/managerReducer/licenseReducer';
 import rbac from '../resources/rbac';
 import versions from '../resources/versions';
 import licenses from '../resources/licenses';
